Fix for bug #694967 - MuPDF crashes after pressing the Back Button while it renders a specific PDF
Some async tasks were still trying to access the globals pointer, which had
been nulled during shutdown, leading to a crash.

Check for NULL and return cleanly in this case.

Possibly we should also be calling AsyncTask.get() to wait for the tasks to
finish, before nulling the global pointer?