Add proper support for when a combobox widget has options that are 2-element arrays
The list box and combo box can have values that are 2-element arrays.  The first element
is the "export" value and the second element is the value that should be shown
in the list box UI.  This fix ensures that we get the proper value to show in the UI.
Also, it adds the option to get the export values.   These are needed if you wish to
update the field dictionary's V (value) entry, which is the currently selected values(s).
This fix works well with gsview.  The other viewers will now display the proper content
in their UI, (unlike before this fix) but may need a bit more work to ensure that the
proper V (value) is updated with changes in the selections.  In addition, we add
selection rectangles to the selected list box items.