Bug 696662: Android viewer: data from URI stream fix.
Apparently, if MuPDF is invoked on a content stream that
comes from a URI, then is.available() can report 0 bytes
(as there is no data buffered). Use a modified formulation
that reads as much data as possible from the stream into
a BufferedOutputStream and then makes a byte array from that.

Would be nicer if the core could actually read from the
stream directly, perhaps, but that can wait for the new JNI
based version.

Thanks to Marc K for identifying the problem and supplying
the patch.