Update JNI code to load 32 or 64 bit DLL as appropriate.
Make the JNI code detect whether it is running on a 32 or
64 bit machine, and change the name of the DLL appropriately.

Update Android Makefile to make mupdf_java32 instead of mupdf_java.

Update Java Makefile to make mupdf_java32 or mupdf_java64 based
on the system it is running on. This choice can be overruled by
defining BITS to be "32" or "64" before calling make.

Update Windows Solution to make mupdf_java32 or mupdf_java64 as
appropriate.